+++ JustSaying.AwsTools.UnitTests/MessageHandling/SqsNotificationListener/Support/ThrowingDuringMessageProcessingStrategy.cs
@@ -3,32 +3,32 @@ using System.Threading.Tasks;
 using JustSaying.Messaging.MessageProcessingStrategies;
 using JustSaying.TestingFramework;
 
-namespace JustSaying.AwsTools.UnitTests.MessageHandling.SqsNotificationListener
+namespace JustSaying.AwsTools.UnitTests.MessageHandling.SqsNotificationListener.Support
 {
-    public class ThrowingBeforeMessageProcessingStrategy : IMessageProcessingStrategy
+    public class ThrowingDuringMessageProcessingStrategy : IMessageProcessingStrategy
     {
         public int MaxBatchSize { get { return int.MaxValue; } }
 
         private readonly TaskCompletionSource<object> _doneSignal;
 
-        public ThrowingBeforeMessageProcessingStrategy(TaskCompletionSource<object> doneSignal)
+        public ThrowingDuringMessageProcessingStrategy(TaskCompletionSource<object> doneSignal)
         {
             _doneSignal = doneSignal;
         }
 
-        public Task BeforeGettingMoreMessages()
+        public async Task BeforeGettingMoreMessages()
         {
-            Fail();
-            return null;
+            await Task.Delay(10);
         }
 
         public void ProcessMessage(Action action)
         {
+            Fail();
         }
 
         private void Fail()
         {
-            Patiently.DelaySendDone(_doneSignal);
+            Tasks.DelaySendDone(_doneSignal);
             throw new Exception("Thrown by test ProcessMessage");
         }
