+++ JustSaying.AwsTools.UnitTests/MessageHandling/SqsNotificationListener/WhenMessageProcessingThrowsDuring.cs
@@ -1,6 +1,7 @@
 using System.Threading.Tasks;
 using Amazon.SQS.Model;
 using JustBehave;
+using JustSaying.AwsTools.UnitTests.MessageHandling.SqsNotificationListener.Support;
 using JustSaying.TestingFramework;
 using NSubstitute;
