+++ JustSaying.TestingFramework/Patiently.cs
@@ -6,33 +6,6 @@ namespace JustSaying.TestingFramework
 {
     public static class Patiently
     {
-        private const int DefaultTimeoutMillis = 1000;
-
-        public static async Task WaitWithTimeoutAsync(Task task)
-        {
-            await WaitWithTimeoutAsync(task, TimeSpan.FromMilliseconds(DefaultTimeoutMillis));
-        }
-
-        public static void DelaySendDone(TaskCompletionSource<object> doneSignal)
-        {
-            Task.Run(async () =>
-            {
-                await Task.Delay(100);
-                doneSignal.SetResult(null);
-            });
-        }
-
-        public static async Task WaitWithTimeoutAsync(Task task, TimeSpan timeoutDuration)
-        {
-            var timeoutTask = Task.Delay(timeoutDuration);
-            var firstToComplete = await Task.WhenAny(task, timeoutTask);
-
-            if (firstToComplete == timeoutTask)
-            {
-                throw new Exception("Task did not complete before timeout of " + timeoutDuration);
-            }
-        }
-
         public static async Task VerifyExpectationAsync(Action expression)
         {
             await VerifyExpectationAsync(expression, 5.Seconds());
